feat: improve pm list
