+++ packages/core/server/src/plugin-manager/utils.ts
@@ -33,6 +33,9 @@ import {
 import deps from './deps';
 import { PluginManagerRepository } from './plugin-manager-repository';
 import { PluginData } from './types';
+import Application from '../application';
+import { findBuiltInPlugins, findLocalPlugins } from './findPackageNames';
+import PluginManager from './plugin-manager';
 
 /**
  * get temp dir
@@ -600,4 +603,34 @@ export async function getPluginBasePath(packageName: string) {
     // skip
   }
   return path.dirname(path.dirname(file));
+}
+
+export async function pmListSummary(app: Application) {
+  const plugins1 = await findBuiltInPlugins();
+  const plugins2 = await findLocalPlugins();
+  let enabledPlugins = [];
+  try {
+    enabledPlugins = (
+      await app.pm.repository.find({
+        filter: {
+          enabled: true,
+        },
+      })
+    ).map((item) => item.packageName);
+  } catch (error) {
+    // ignore error
+  }
+  const items = await Promise.all(
+    [...plugins1, ...plugins2].map(async (name) => {
+      const item = await PluginManager.parseName(name);
+      const json = await PluginManager.getPackageJson(item.packageName);
+      return {
+        displayName: json.displayName || name,
+        packageName: item.packageName,
+        enabled: enabledPlugins.includes(item.packageName),
+        description: json.description,
+      };
+    }),
+  );
+  return items;
 }
